Add World Manager with time and weather mechanics (#7)
* Add simple day and night cycle with day counter

When time passes, day and night states are cycled, and a counter
keeps track of the current day so that it can be shown to the
player in the future.

Part of the World Manager task

* Add weighted random weather system

Adds a weather system with 4 possible states, which are chosen at
random each passing time of day with weights so to prioritize clear
weather and make a catastrophical weather condition very rare.

Random system is not yet ideal.

Part of the World Manager task.

* Add day counter with time icon and time pass button

Integrates the World Manager system with the game via UI using a
day counter, an icon for the time of day, and a button to pass time.

Also adds a toString method to the World Manager.

Part of the World Manager task

* Add color modulation for night time and shop restock every day

The shop should restock every day, and with the world manager we
can now make this a reality.

Since there is a day time and a night time, the difference is shown
with different color modulations.

Part of the World Manager task.

* Integrate weather system with time cycle color modulation

Adds the weather system to the UI and the color modulation system
implemented for the day/night cycle. The weather color is mixed
with the time color for the modulation. A weather icon was added
to the World UI.

Finishes the World Manager task.