Plant index (#21)
* Add Plant Index stub

The Plant Index is the player's goal, to collect them all. The UI
part of the index is added as a stub here.

Part of the Plant Index task.

* Add Plant Index back end

Add the actual plant goal back end and use the index to list the
objective. Handle collecting plants.

* Save new plant name when changed by user

The plant name was so far editable but would not save the changes.
Now the changes are saved right when the text is changed.

Finishes the Plant Index basic task.